+++ test/config.php
@@ -1,6 +1,6 @@
 <?php
 
-require_once(__DIR__.'/../src/Amiss.php');
+require_once __DIR__.'/../vendor/autoload.php';
 Amiss::register();
 
 date_default_timezone_set('Australia/Melbourne');